backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
google-generativeai==0.3.2
supabase==2.1.0
python-dotenv==1.0.0
httpx
pydantic